+++ src/PostgREST/Error.hs
@@ -52,74 +52,107 @@ class (JSON.ToJSON a) => PgrstError a where
 
 
 
+
 data ApiRequestError
-  = ActionInappropriate
-  | InvalidRange
+  = ParseRequestError Text Text
+  | ActionInappropriate
   | InvalidBody ByteString
-  | ParseRequestError Text Text
-  | NoRelBetween Text Text
-  | AmbiguousRelBetween Text Text [Relationship]
-  | AmbiguousRpc [ProcDescription]
-  | NoRpc Text Text [Text] Bool ContentType Bool
+  | InvalidRange
+  | UnsupportedVerb                -- Unreachable?
   | InvalidFilters
   | UnacceptableSchema [Text]
   | ContentTypeError [ByteString]
-  | UnsupportedVerb                -- Unreachable?
+  | NoRelBetween Text Text
+  | AmbiguousRelBetween Text Text [Relationship]
+  | NoRpc Text Text [Text] Bool ContentType Bool  
+  | AmbiguousRpc [ProcDescription]  
 
 instance PgrstError ApiRequestError where
-  status InvalidRange            = HT.status416
-  status InvalidFilters          = HT.status405
-  status (InvalidBody _)         = HT.status400
-  status UnsupportedVerb         = HT.status405
-  status ActionInappropriate     = HT.status405
   status (ParseRequestError _ _) = HT.status400
-  status (NoRelBetween _ _)      = HT.status400
-  status AmbiguousRelBetween{}   = HT.status300
-  status (AmbiguousRpc _)        = HT.status300
-  status NoRpc{}                 = HT.status404
+  status ActionInappropriate     = HT.status405
+  status (InvalidBody _)         = HT.status400
+  status InvalidRange            = HT.status416
+  status UnsupportedVerb         = HT.status405
+  status InvalidFilters          = HT.status405
   status (UnacceptableSchema _)  = HT.status406
   status (ContentTypeError _)    = HT.status415
+  status (NoRelBetween _ _)      = HT.status400
+  status AmbiguousRelBetween{}   = HT.status300
+  status NoRpc{}                 = HT.status404
+  status (AmbiguousRpc _)        = HT.status300
 
   headers _ = [ContentType.toHeader CTApplicationJSON]
 
 instance JSON.ToJSON ApiRequestError where
   toJSON (ParseRequestError message details) = JSON.object [
-    "message" .= message, "details" .= details]
+    "code"    .= ApiRequestErrorCode00,
+    "message" .= message,
+    "details" .= details,
+    "hint"    .= JSON.Null]
   toJSON ActionInappropriate = JSON.object [
-    "message" .= ("Bad Request" :: Text)]
+    "code"    .= ApiRequestErrorCode01,
+    "message" .= ("Bad Request" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
   toJSON (InvalidBody errorMessage) = JSON.object [
-    "message" .= (toS errorMessage :: Text)]
+    "code"    .= ApiRequestErrorCode02,
+    "message" .= (toS errorMessage :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
   toJSON InvalidRange = JSON.object [
-    "message" .= ("HTTP Range error" :: Text)]
+    "code"    .= ApiRequestErrorCode03,
+    "message" .= ("HTTP Range error" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON UnsupportedVerb = JSON.object [
+    "code"    .= ApiRequestErrorCode04,
+    "message" .= ("Unsupported HTTP verb" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON InvalidFilters = JSON.object [
+    "code"    .= ApiRequestErrorCode05,
+    "message" .= ("Filters must include all and only primary key columns with 'eq' operators" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON (UnacceptableSchema schemas) = JSON.object [
+    "code"    .= ApiRequestErrorCode06,
+    "message" .= ("The schema must be one of the following: " <> T.intercalate ", " schemas),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON (ContentTypeError cts)    = JSON.object [
+    "code"    .= ApiRequestErrorCode07,
+    "message" .= ("None of these Content-Types are available: " <> (toS . intercalate ", " . map toS) cts :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+
   toJSON (NoRelBetween parent child) = JSON.object [
-    "hint"    .= ("If a new foreign key between these entities was created in the database, try reloading the schema cache." :: Text),
-    "message" .= ("Could not find a relationship between " <> parent <> " and " <> child <> " in the schema cache" :: Text)]
+    "code"    .= SchemaCacheErrorCode01,
+    "message" .= ("Could not find a relationship between " <> parent <> " and " <> child <> " in the schema cache" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= ("If a new foreign key between these entities was created in the database, try reloading the schema cache." :: Text)]
   toJSON (AmbiguousRelBetween parent child rels) = JSON.object [
-    "hint"    .= ("By following the 'details' key, disambiguate the request by changing the url to /origin?select=relationship(*) or /origin?select=target!relationship(*)" :: Text),
+    "code"    .= SchemaCacheErrorCode02,
     "message" .= ("More than one relationship was found for " <> parent <> " and " <> child :: Text),
-    "details" .= (compressedRel <$> rels) ]
-  toJSON (AmbiguousRpc procs)  = JSON.object [
-    "hint"    .= ("Try renaming the parameters or the function itself in the database so function overloading can be resolved" :: Text),
-    "message" .= ("Could not choose the best candidate function between: " <> T.intercalate ", " [pdSchema p <> "." <> pdName p <> "(" <> T.intercalate ", " [ppName a <> " => " <> ppType a | a <- pdParams p] <> ")" | p <- procs])]
+    "details" .= (compressedRel <$> rels),
+    "hint"    .= ("By following the 'details' key, disambiguate the request by changing the url to /origin?select=relationship(*) or /origin?select=target!relationship(*)" :: Text)]  
   toJSON (NoRpc schema procName argumentKeys hasPreferSingleObject contentType isInvPost)  =
     let prms = "(" <> T.intercalate ", " argumentKeys <> ")" in JSON.object [
-    "hint"    .= ("If a new function was created in the database with this name and parameters, try reloading the schema cache." :: Text),
+    "code"    .= SchemaCacheErrorCode04,
     "message" .= ("Could not find the " <> schema <> "." <> procName <>
-      (case (hasPreferSingleObject, isInvPost, contentType) of
-        (True, _, _)                 -> " function with a single json or jsonb parameter"
-        (_, True, CTTextPlain)       -> " function with a single unnamed text parameter"
-        (_, True, CTOctetStream)     -> " function with a single unnamed bytea parameter"
-        (_, True, CTApplicationJSON) -> prms <> " function or the " <> schema <> "." <> procName <>" function with a single unnamed json or jsonb parameter"
-        _                            -> prms <> " function") <>
-      " in the schema cache")]
-  toJSON UnsupportedVerb = JSON.object [
-    "message" .= ("Unsupported HTTP verb" :: Text)]
-  toJSON InvalidFilters = JSON.object [
-    "message" .= ("Filters must include all and only primary key columns with 'eq' operators" :: Text)]
-  toJSON (UnacceptableSchema schemas) = JSON.object [
-    "message" .= ("The schema must be one of the following: " <> T.intercalate ", " schemas)]
-  toJSON (ContentTypeError cts)    = JSON.object [
-    "message" .= ("None of these Content-Types are available: " <> (toS . intercalate ", " . map toS) cts :: Text)]
+          (case (hasPreferSingleObject, isInvPost, contentType) of
+            (True, _, _)                 -> " function with a single json or jsonb parameter"
+            (_, True, CTTextPlain)       -> " function with a single unnamed text parameter"
+            (_, True, CTOctetStream)     -> " function with a single unnamed bytea parameter"
+            (_, True, CTApplicationJSON) -> prms <> " function or the " <> schema <> "." <> procName <>" function with a single unnamed json or jsonb parameter"
+            _                            -> prms <> " function") <>
+          " in the schema cache"),
+    "details" .= JSON.Null,
+    "hint"    .= ("If a new function was created in the database with this name and parameters, try reloading the schema cache." :: Text)]
+  toJSON (AmbiguousRpc procs)  = JSON.object [
+    "code"    .= SchemaCacheErrorCode03,
+    "message" .= ("Could not choose the best candidate function between: " <> T.intercalate ", " [pdSchema p <> "." <> pdName p <> "(" <> T.intercalate ", " [ppName a <> " => " <> ppType a | a <- pdParams p] <> ")" | p <- procs]),
+    "details" .= JSON.Null,
+    "hint"    .= ("Try renaming the parameters or the function itself in the database so function overloading can be resolved" :: Text)]
 
 compressedRel :: Relationship -> JSON.Value
 compressedRel Relationship{..} =
@@ -161,46 +194,52 @@ instance JSON.ToJSON PgError where
 
 instance JSON.ToJSON P.UsageError where
   toJSON (P.ConnectionError e) = JSON.object [
-    "code"    .= ("" :: Text),
+    "code"    .= ConnectionErrorCode00,
     "message" .= ("Database connection error. Retrying the connection." :: Text),
-    "details" .= (toSL $ fromMaybe "" e :: Text)]
+    "details" .= (toSL $ fromMaybe "" e :: Text),
+    "hint"    .= JSON.Null]
   toJSON (P.SessionError e) = JSON.toJSON e -- H.Error
 
 instance JSON.ToJSON H.QueryError where
   toJSON (H.QueryError _ _ e) = JSON.toJSON e
 
 instance JSON.ToJSON H.CommandError where
-  toJSON (H.ResultError (H.ServerError c m d h)) = case toS c of
-    'P':'T':_ -> JSON.object [
-        "details" .= (fmap toS d :: Maybe Text),
-        "hint"    .= (fmap toS h :: Maybe Text)]
-
-    _         -> JSON.object [
+  toJSON (H.ResultError (H.ServerError c m d h)) = JSON.object [
         "code"    .= (toS c      :: Text),
         "message" .= (toS m      :: Text),
         "details" .= (fmap toS d :: Maybe Text),
         "hint"    .= (fmap toS h :: Maybe Text)]
 
   toJSON (H.ResultError (H.UnexpectedResult m)) = JSON.object [
-    "message" .= (m :: Text)]
+    "code"    .= HasqlErrorCode01,
+    "message" .= (m :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
   toJSON (H.ResultError (H.RowError i H.EndOfInput)) = JSON.object [
+    "code"    .= HasqlErrorCode02,
     "message" .= ("Row error: end of input" :: Text),
     "details" .= ("Attempt to parse more columns than there are in the result" :: Text),
     "hint"    .= (("Row number " <> show i) :: Text)]
   toJSON (H.ResultError (H.RowError i H.UnexpectedNull)) = JSON.object [
+    "code"    .= HasqlErrorCode03,
     "message" .= ("Row error: unexpected null" :: Text),
     "details" .= ("Attempt to parse a NULL as some value." :: Text),
     "hint"    .= (("Row number " <> show i) :: Text)]
   toJSON (H.ResultError (H.RowError i (H.ValueError d))) = JSON.object [
+    "code"    .= HasqlErrorCode04,
     "message" .= ("Row error: Wrong value parser used" :: Text),
     "details" .= d,
     "hint"    .= (("Row number " <> show i) :: Text)]
   toJSON (H.ResultError (H.UnexpectedAmountOfRows i)) = JSON.object [
+    "code"    .= HasqlErrorCode05,
     "message" .= ("Unexpected amount of rows" :: Text),
-    "details" .= i]
+    "details" .= i,
+    "hint"    .= JSON.Null]
   toJSON (H.ClientError d) = JSON.object [
+    "code"    .= ConnectionErrorCode01,
     "message" .= ("Database client error. Retrying the connection." :: Text),
-    "details" .= (fmap toS d :: Maybe Text)]
+    "details" .= (fmap toS d :: Maybe Text),
+    "hint"    .= JSON.Null]
 
 pgErrorStatus :: Bool -> P.UsageError -> HT.Status
 pgErrorStatus _      (P.ConnectionError _)                                      = HT.status503
@@ -267,28 +306,28 @@ checkIsFatal _ = Nothing
 
 
 data Error
-  = GucHeadersError
-  | GucStatusError
-  | BinaryFieldError ContentType
-  | ConnectionLostError
-  | PutMatchingPkError
-  | PutRangeNotAllowedError
+  = ConnectionLostError
   | JwtTokenMissing
   | JwtTokenInvalid Text
+  | GucHeadersError
+  | GucStatusError
+  | BinaryFieldError ContentType
+  | PutMatchingPkError
+  | PutRangeNotAllowedError
   | SingularityError Integer
   | NotFound
-  | ApiRequestError ApiRequestError
   | PgErr PgError
+  | ApiRequestError ApiRequestError
 
 instance PgrstError Error where
+  status ConnectionLostError     = HT.status503
+  status JwtTokenMissing         = HT.status500
+  status (JwtTokenInvalid _)     = HT.unauthorized401
   status GucHeadersError         = HT.status500
   status GucStatusError          = HT.status500
   status (BinaryFieldError _)    = HT.status406
-  status ConnectionLostError     = HT.status503
   status PutMatchingPkError      = HT.status400
   status PutRangeNotAllowedError = HT.status400
-  status JwtTokenMissing         = HT.status500
-  status (JwtTokenInvalid _)     = HT.unauthorized401
   status (SingularityError _)    = HT.status406
   status NotFound                = HT.status404
   status (PgErr err)             = status err
@@ -301,28 +340,56 @@ instance PgrstError Error where
   headers _                        = [ContentType.toHeader CTApplicationJSON]
 
 instance JSON.ToJSON Error where
-  toJSON GucHeadersError           = JSON.object [
-    "message" .= ("response.headers guc must be a JSON array composed of objects with a single key and a string value" :: Text)]
-  toJSON GucStatusError           = JSON.object [
-    "message" .= ("response.status guc must be a valid status code" :: Text)]
-  toJSON (BinaryFieldError ct)          = JSON.object [
-    "message" .= ((toS (ContentType.toMime ct) <> " requested but more than one column was selected") :: Text)]
   toJSON ConnectionLostError       = JSON.object [
-    "message" .= ("Database connection lost. Retrying the connection." :: Text)]
-
-  toJSON PutRangeNotAllowedError   = JSON.object [
-    "message" .= ("Range header and limit/offset querystring parameters are not allowed for PUT" :: Text)]
-  toJSON PutMatchingPkError        = JSON.object [
-    "message" .= ("Payload values do not match URL in primary key column(s)" :: Text)]
-
-  toJSON (SingularityError n)      = JSON.object [
-    "message" .= ("JSON object requested, multiple (or no) rows returned" :: Text),
-    "details" .= T.unwords ["Results contain", show n, "rows,", toS (ContentType.toMime CTSingularJSON), "requires 1 row"]]
+      "code"    .= ConnectionErrorCode02,
+      "message" .= ("Database connection lost. Retrying the connection." :: Text),
+      "details" .= JSON.Null,
+      "hint"    .= JSON.Null]
 
   toJSON JwtTokenMissing           = JSON.object [
-    "message" .= ("Server lacks JWT secret" :: Text)]
+      "code"    .= JWTErrorCode00,
+      "message" .= ("Server lacks JWT secret" :: Text),
+      "details" .= JSON.Null,
+      "hint"    .= JSON.Null]
   toJSON (JwtTokenInvalid message) = JSON.object [
-    "message" .= (message :: Text)]
+      "code"    .= JWTErrorCode01,
+      "message" .= (message :: Text),
+      "details" .= JSON.Null,
+      "hint"    .= JSON.Null]
+
+  toJSON GucHeadersError           = JSON.object [
+    "code"    .= GeneralErrorCode01,
+    "message" .= ("response.headers guc must be a JSON array composed of objects with a single key and a string value" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON GucStatusError           = JSON.object [
+    "code"    .= GeneralErrorCode02,
+    "message" .= ("response.status guc must be a valid status code" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON (BinaryFieldError ct)          = JSON.object [
+    "code"    .= GeneralErrorCode03,
+    "message" .= ((toS (ContentType.toMime ct) <> " requested but more than one column was selected") :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+
+  toJSON PutRangeNotAllowedError   = JSON.object [
+    "code"    .= GeneralErrorCode04,
+    "message" .= ("Range header and limit/offset querystring parameters are not allowed for PUT" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+  toJSON PutMatchingPkError        = JSON.object [
+    "code"    .= GeneralErrorCode05,
+    "message" .= ("Payload values do not match URL in primary key column(s)" :: Text),
+    "details" .= JSON.Null,
+    "hint"    .= JSON.Null]
+
+  toJSON (SingularityError n)      = JSON.object [
+    "code"    .= GeneralErrorCode06,
+    "message" .= ("JSON object requested, multiple (or no) rows returned" :: Text),
+    "details" .= T.unwords ["Results contain", show n, "rows,", toS (ContentType.toMime CTSingularJSON), "requires 1 row"],
+    "hint"    .= JSON.Null]
+
   toJSON NotFound = JSON.object []
   toJSON (PgErr err) = JSON.toJSON err
   toJSON (ApiRequestError err) = JSON.toJSON err
@@ -333,3 +400,85 @@ invalidTokenHeader m =
 
 singularityError :: (Integral a) => a -> Error
 singularityError = SingularityError . toInteger
+
+-- Error codes are defined following the order of the process:
+-- Connection -> API Requests -> Auth -> DB Queries
+data ErrorCode
+  -- PostgreSQL connection errors
+  = ConnectionErrorCode00
+  | ConnectionErrorCode01
+  | ConnectionErrorCode02
+  -- API Request errors
+  | ApiRequestErrorCode00
+  | ApiRequestErrorCode01
+  | ApiRequestErrorCode02
+  | ApiRequestErrorCode03
+  | ApiRequestErrorCode04
+  | ApiRequestErrorCode05
+  | ApiRequestErrorCode06
+  | ApiRequestErrorCode07
+  -- Schema Cache errors
+  | SchemaCacheErrorCode01
+  | SchemaCacheErrorCode02
+  | SchemaCacheErrorCode03
+  | SchemaCacheErrorCode04
+  -- JWT authentication errors
+  | JWTErrorCode00
+  | JWTErrorCode01
+  -- Hasql library errors
+  | HasqlErrorCode00
+  | HasqlErrorCode01
+  | HasqlErrorCode02
+  | HasqlErrorCode03
+  | HasqlErrorCode04
+  | HasqlErrorCode05
+  -- Uncategorized errors that are not related to a single module
+  | GeneralErrorCode00
+  | GeneralErrorCode01
+  | GeneralErrorCode02
+  | GeneralErrorCode03
+  | GeneralErrorCode04
+  | GeneralErrorCode05
+  | GeneralErrorCode06
+
+-- The prefixes should go according to the process order:
+instance JSON.ToJSON ErrorCode where
+  toJSON e = JSON.toJSON (buildErrorCode e)
+
+buildErrorCode :: ErrorCode -> Text
+buildErrorCode code = "PGRST" <> case code of
+  ConnectionErrorCode00  -> "000"
+  ConnectionErrorCode01  -> "001"
+  ConnectionErrorCode02  -> "002"
+
+  ApiRequestErrorCode00  -> "100"
+  ApiRequestErrorCode01  -> "101"
+  ApiRequestErrorCode02  -> "102"
+  ApiRequestErrorCode03  -> "103"
+  ApiRequestErrorCode04  -> "104"
+  ApiRequestErrorCode05  -> "105"
+  ApiRequestErrorCode06  -> "106"
+  ApiRequestErrorCode07  -> "107"
+  
+  SchemaCacheErrorCode01 -> "200"
+  SchemaCacheErrorCode02 -> "201"
+  SchemaCacheErrorCode03 -> "202"
+  SchemaCacheErrorCode04 -> "203"
+
+  JWTErrorCode00         -> "300"
+  JWTErrorCode01         -> "301"
+
+  HasqlErrorCode00       -> "400"
+  HasqlErrorCode01       -> "401"
+  HasqlErrorCode02       -> "402"
+  HasqlErrorCode03       -> "403"
+  HasqlErrorCode04       -> "404"
+  HasqlErrorCode05       -> "405"
+
+  GeneralErrorCode00     -> "500"
+  GeneralErrorCode01     -> "501"
+  GeneralErrorCode02     -> "502"
+  GeneralErrorCode03     -> "503"
+  GeneralErrorCode04     -> "504"
+  GeneralErrorCode05     -> "505"
+  GeneralErrorCode06     -> "506"